% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_var_by_row}
\alias{calc_var_by_row}
\title{Calculate Variance by Row}
\usage{
calc_var_by_row(x, ...)
}
\arguments{
\item{x}{matrix or data frame}

\item{...}{additional paramaters addressed to rowSums or rowMeans}
}
\description{
Calculate variance by row in a matrix or data frame
}
