% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{draw_colnames_45}
\alias{draw_colnames_45}
\title{plot aux function}
\usage{
draw_colnames_45(coln, gaps, ...)
}
\arguments{
\item{coln}{number of columns}

\item{gaps}{gaps}

\item{...}{additional parameters}
}
\description{
Auxiliar function for rotating plot_heatmap x axis
}
