% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gsea_analysis}
\alias{gsea_analysis}
\title{Perform Gene Set Enrichment Analysis}
\usage{
gsea_analysis(de_res, gene_sets, file_ext = "gmt")
}
\arguments{
\item{de_res}{results from de_analysis function}

\item{gene_sets}{tidy data frame of the gene sets or
named list of genes named by gene set}

\item{file_ext}{File extension where the gene
sets should be loaded GMT is the default for GSEA
rds to load rds saved files. DEPRECATED}
}
\value{
A \code{tibble} containing GSEA results
}
\description{
Perform GSEA analysis.
}
