% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe, as in Pipeline}
\arguments{
\item{lhs, rhs}{An object and a function to apply to it}
}
\description{
Exported from dplyr, use \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of
abs(nrow(head(mtcars)))
# you can write
mtcars \%>\% head() \%>\% nrow() \%>\% abs()
}
