% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gene_abundance}
\alias{plot_gene_abundance}
\title{Plot Gene Expression}
\usage{
plot_gene_abundance(gene, tx, sample_table = NULL, color_by = NULL,
  filter = NULL)
}
\arguments{
\item{gene}{gene or vector of genes}

\item{tx}{\code{txi} object, object imported with
`import_tx()` or data frame wih gene abundance}

\item{sample_table}{metadata data frame, containing description
of each sample and experimental design}

\item{color_by}{variable from sample table to group samples in the plot}

\item{filter}{samples to filter, dafault = NULL}
}
\description{
Plot gene expression points
}
