% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot GSEA Enrichment Curve and Table}
\usage{
plot_gsea(gsea_res, de_res, gene_sets, top_sets = 10,
  enrichment = FALSE, title = NULL)
}
\arguments{
\item{gsea_res}{result from gsea_analysis}

\item{de_res}{results from de_analysis function}

\item{gene_sets}{named list of genes named by gene set}

\item{top_sets}{Number of sets to be plotted in the gsea table,
the number of sets to be plotted is the double of the value
of top_sets
or a string for individual gene set for enrichment plot}

\item{enrichment}{If set to TRUE, enrichment for individual
gene set will be plotted.
The default value is FALSE, for plotting GSEA table.}

\item{title}{title label to be shown for enchment plot,
should be a character vector of the text to be plotted
or TRUE for using the .default = NULL, no text plotted.}
}
\value{
graphs for gsea table or gsea individual set
}
\description{
Plots enrichment table and individual enrichment graph for
GSEA analysis results
}
