% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gsea_heatmap}
\alias{plot_gsea_heatmap}
\title{Plot a Heatmap From gsea_analysis}
\usage{
plot_gsea_heatmap(..., pvalue = 0.05, fdr = FALSE, names = NULL,
  occurrence = NULL)
}
\arguments{
\item{...}{list of gsea_analisys results}

\item{pvalue}{p-value cut-off for enrichment or differential consideration,
default = 0.05}

\item{fdr}{logical, default = FALSE should False discovery rate adjusted
pvalue be used instead of pvalue}

\item{names}{vector with names to be used at the x axis label}

\item{occurrence}{Number of samples that the gene set
needs to be enriched}
}
\description{
Plot a heatmap of enriched gene sets from gsea_analysis
}
