% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gsea_res}
\alias{plot_gsea_res}
\title{Plot GSEA Results and Leading Edge}
\usage{
plot_gsea_res(gsea_res, pvalue = 0.05, fdr = FALSE)
}
\arguments{
\item{gsea_res}{result from gsea_analysis}

\item{pvalue}{p-value cut-off for enrichment or differential consideration,
default = 0.05}

\item{fdr}{logical, default = FALSE should False discovery rate adjusted
pvalue be used instead of pvalue}
}
\description{
plot gsea results with leading edge genes for enriched gene sets
}
