% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_le_heatmap}
\alias{plot_le_heatmap}
\title{Plot Leading Edge Heatmap}
\usage{
plot_le_heatmap(gsea_res, pvalue = 0.05, fdr = FALSE, num = NULL)
}
\arguments{
\item{gsea_res}{result from gsea_analysis}

\item{pvalue}{p-value cut-off for enrichment or differential consideration,
default = 0.05}

\item{fdr}{logical, default = FALSE should False discovery rate adjusted
pvalue be used instead of pvalue}

\item{num}{number of genes to plot, based on the leading edge occurence,
if num is negative, it is used the absolute num of genes with
lesser occurence, if num is NULL all leading edge elements
are used. Default = NULL.}
}
\description{
plot Ledading Edge analysis results enriched gene sets
}
