% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot Two Dimensional PCA}
\usage{
plot_pca(tx, sample_table, color_by = NULL, num = NULL)
}
\arguments{
\item{tx}{txi object, matrix or data frame with trancript abundance}

\item{sample_table}{metadata data frame}

\item{color_by}{variable from sample table to group samples in the plot}

\item{num}{number of genes to plot, filtering by variance between samples}
}
\description{
Plot the results of a two dimensional principal component
analysis.
}
