% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_venn_diagram}
\alias{plot_venn_diagram}
\title{Plot Venn Diagram}
\usage{
plot_venn_diagram(..., pvalue = 0.05, fdr = FALSE, names = NULL,
  filename, width = 9, height = NULL)
}
\arguments{
\item{...}{de_analysis or gsea_analysis data frames}

\item{pvalue}{p-value cut-off for enrichment or differential consideration,
default = 0.05}

\item{fdr}{logical, default = FALSE should False discovery rate adjusted
pvalue be used instead of pvalue}

\item{names}{vector with names to be used for each group}

\item{filename}{path to were the image should be saved}

\item{width}{value to be used as the width of the plot, the venn diagram
will be plotted as a square with side of the smaller
value, between width and height}

\item{height}{value to be used as the height of the plot, the venn diagram
will be plotted as a square of the smaller value, between
width and height. heigth cam be ommited and the plot will be
a square.}
}
\description{
Plot Venn diagrams, from de_analysis or gsea_analysis results
**Currentry, this version only plots venn diagram with four groups and just
save files to file, don't plot directly**
}
