% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(de_res, alpha = 0.05, lfc_threshold = NULL, fdr = FALSE)
}
\arguments{
\item{de_res}{de_analysis result}

\item{alpha}{default = 0.05}

\item{lfc_threshold}{absolute value of log2FoldChange to limit the significance}

\item{fdr}{logical, should consider False discovery rate
default = FALSE}
}
\value{
A \code{ggplot} object
}
\description{
A volcano plot from DE genes analysis results.
}
