% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_tx_mat}
\alias{prepare_tx_mat}
\title{plot aux function 2}
\usage{
prepare_tx_mat(tx)
}
\arguments{
\item{tx}{expression data or result from import_tx}
}
\description{
Prepare data for plotting PCA or Expression heatmap
}
