% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_deg}
\alias{retrieve_deg}
\title{Retrieve Differentially Expressed Genes}
\usage{
retrieve_deg(de_res, alpha = 0.05, fdr = FALSE)
}
\arguments{
\item{de_res}{de_analysis result}

\item{alpha}{default = 0.05}

\item{fdr}{logical, should consider False discovery rate
default = FALSE}
}
\value{
a \code{vector} of gene names
}
\description{
Extract differentially expressed genes from
DESeq2 results based on p-value cuttof.
}
