% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_deg_table}
\alias{retrieve_deg_table}
\title{Differential Expression results}
\usage{
retrieve_deg_table(..., pvalue = 0.05, fdr = FALSE, names = NULL)
}
\arguments{
\item{...}{de_analysis or gsea_analysis results data frames}

\item{pvalue}{p-value cut-off for enrichment or differential consideration,
default = 0.05}

\item{fdr}{logical, default = FALSE should False discovery rate adjusted
pvalue be used instead of pvalue}

\item{names}{vector with names to be used for each group}
}
\description{
Filter upregulated and downregulated genes from
Differential expression analysis in a table
}
