% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_mapping_rate}
\alias{retrieve_mapping_rate}
\title{Retrieve Mapping Rate}
\usage{
retrieve_mapping_rate(path)
}
\arguments{
\item{path}{path to Salmon results files}
}
\value{
A \code{tibble} object
}
\description{
Retrieve total number of fragments,
total mapped reads and
mapping rate from a Salmon mapped
sample.
}
