% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{salmon_libtype}
\alias{salmon_libtype}
\title{salmon_libtype}
\usage{
salmon_libtype(dir)
}
\arguments{
\item{dir}{path to Salmon results files}
}
\value{
A \code{data_frame} object
}
\description{
retrieve optimal library preparation
protocol used for Salmon method of
mapping
}
\details{
Function used to import transcripts abundance data
}
