% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{txomics-package}
\alias{txomics}
\alias{txomics-package}
\title{txomics: Functional Analysis and Visualization of Transcriptomic Data}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Functional analysis of transcriptomic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://luciorq.github.io/txomics}
  \item \url{https://github.com/luciorq/txomics}
  \item Report bugs at \url{https://github.com/luciorq/txomics/issues}
}

}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com} (0000-0002-6090-1834)

}
\keyword{internal}
