% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{txomics}
\alias{txomics}
\alias{txomics-package}
\title{\code{txomics} package}
\description{
Functional analysis of transcriptomic data
}
\details{
See the README on
\href{https://github.com/luciorq/txomics#readme}{GitHub}
}
