% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{de_analysis}
\alias{de_analysis}
\title{Perform Differential Expression Analysis}
\usage{
de_analysis(txi, sample_table, contrast_var, numerator, denominator,
  batch_var = NULL, beta_prior = TRUE, force_rep = FALSE)
}
\arguments{
\item{txi}{tximport object containing reads counts}

\item{sample_table}{table containing experimental design
and sample info}

\item{contrast_var}{column name from the sample_table}

\item{numerator}{factor from contrast_var to be considered as treatment}

\item{denominator}{factor from contrast_var to be considered as control}

\item{batch_var}{column from sample_table to be considered for batch effect}

\item{beta_prior}{default = TRUE, set to FALSE to use apeglm method}

\item{force_rep}{default = FALSE, experimental}
}
\value{
A \code{tibble} object
}
\description{
Run DESeq2 pipeline for differentially expressed genes.
}
