% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GSEA_analysis}
\alias{GSEA_analysis}
\title{GSEA_analysis}
\usage{
GSEA_analysis(DE_res, gene_sets, file_ext = "gmt")
}
\arguments{
\item{DE_res}{DESeqResults object containing DE analysis results}

\item{gene_sets}{named list of genes named by gene set}

\item{file_ext}{File extension where the gene sets should be loaded GMT is the default for GSEA}
}
\value{
A \code{data_frame} containing GSEA results
}
\description{
perform GSEA analysis
}
