% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_tx}
\alias{import_tx}
\title{Import Expression Data}
\usage{
import_tx(dir, source = "salmon", names = "vectorbase")
}
\arguments{
\item{dir}{path to files}

\item{source}{method used}

\item{names}{source of gene names identifier}
}
\value{
A \code{txi} object containing
}
\description{
Function used to import transcripts count and abundance data
}
