% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LE_analysis}
\alias{LE_analysis}
\title{LE_analysis}
\usage{
LE_analysis(GSEA_res, set)
}
\arguments{
\item{GSEA_res}{fgsea result data frame}

\item{set}{gene set to be extracted}
}
\value{
A \code{vector} of leading edge genes for the set
}
\description{
extract Leading Edge genes from GSEA results
}
