% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_deg_heatmap}
\alias{plot_deg_heatmap}
\title{Plot Differentially Expressed Genes Heatmap}
\usage{
plot_deg_heatmap(tx, de_res, pvalue = 0.05, fdr = FALSE)
}
\arguments{
\item{tx}{\code{txi} object, object imported with
`import_tx()` or data frame wih gene abundance}

\item{de_res}{results from de_analysis function}

\item{pvalue}{p-value cut-off for enrichment or differential consideration,
default = 0.05}

\item{fdr}{logical, default = FALSE should False discovery rate adjusted
pvalue be used instead of pvalue}
}
\description{
Plot a expression heatmap of differentially expressed genes de_analysis
}
