% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot Flat Three Dimensional PCA}
\usage{
plot_pca(tx, sample_table, color_by = NULL, num = NULL,
  label = FALSE)
}
\arguments{
\item{tx}{txi object, matrix or data frame with trancript abundance}

\item{sample_table}{metadata data frame, containing description
of each sample and experimental design}

\item{color_by}{variable from sample table to group samples in the plot}

\item{num}{number of genes to plot, filtering by variance between samples,
if num is negative, it is used the absolute num of genes with
smaller, if num is NULL all genes are used.
default = NULL.}

\item{label}{print labels for each sample. Default = FALSE}
}
\description{
Plot the results of a pseudo/flat tri-dimensional
principal component analysis, using 2 two dimensional plots.
}
