% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(DE_res, lfc_threshold = NULL, FDR = FALSE)
}
\arguments{
\item{DE_res}{DESeq2 results object}

\item{lfc_threshold}{absolute value of log2FoldChange to limit the significance}

\item{FDR}{use FDR adjusted p-value, logical}
}
\value{
A \code{ggplot} object
}
\description{
A volcano plot from DE genes analysis results.
}
