% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(de_res, pvalue = 0.05, lfc_threshold = NULL,
  fdr = FALSE)
}
\arguments{
\item{de_res}{de_analysis results}

\item{pvalue}{p-value cut-off for enrichment or differential consideration,
default = 0.05}

\item{lfc_threshold}{absolute value of log2FoldChange to limit the significance}

\item{fdr}{logical, default = FALSE should False discovery rate adjusted
pvalue be used instead of pvalue}
}
\value{
A \code{ggplot} object
}
\description{
A volcano plot from DE genes analysis results.
}
