% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_deg}
\alias{retrieve_deg}
\title{Retrieve Differentially Expressed Genes}
\usage{
retrieve_deg(de_res, pvalue = 0.05, fdr = FALSE)
}
\arguments{
\item{de_res}{de_analysis result}

\item{pvalue}{p-value cut-off for enrichment or differential consideration,
default = 0.05}

\item{fdr}{logical, default = FALSE should False discovery rate adjusted
pvalue be used instead of pvalue}
}
\value{
a \code{vector} of gene names
}
\description{
Extract differentially expressed genes from
DESeq2 results based on p-value cuttof.
}
