% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_DEG}
\alias{retrieve_DEG}
\title{Retrieve Differentially Expressed Genes}
\usage{
retrieve_DEG(DE_res)
}
\arguments{
\item{DE_res}{DESeq2 results object}
}
\value{
a \code{vector} of DEGs
}
\description{
Extract differentially expressed genes from
DESeq2 results based on p-value cuttof.
}
