% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_mapping_rate}
\alias{retrieve_mapping_rate}
\title{retrieve_mapping_rate}
\usage{
retrieve_mapping_rate(dir)
}
\arguments{
\item{dir}{path to Salmon results files}
}
\value{
A \code{data_frame} object
}
\description{
retrieve total number of fragments,
total mapped reads and
mapping rate from a Salmon mapped
sample
}
