% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{salmon_libtype}
\alias{salmon_libtype}
\title{Infer Library Construction Type}
\usage{
salmon_libtype(path)
}
\arguments{
\item{path}{path to Salmon results files}
}
\value{
A \code{tibble} object
}
\description{
Retrieve optimal library preparation
protocol used for Salmon method of
mapping.
}
