% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txomics-package.R
\docType{package}
\name{txomics-package}
\alias{txomics}
\alias{txomics-package}
\title{txomics: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID}) [copyright holder]

}
\keyword{internal}
